Cache diagnostic analyzer computation
